requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
python-dateutil==2.8.2
pandas==1.3.3
matplotlib==3.4.3
seaborn==0.11.2
scikit-learn==1.0.2
numpy==1.21.6
joblib==1.1.0
requests==2.27.1 